brbseq: Support single ended input when creating the fqdemux samplesheet.
For a read structure where the barcode+umi are in R1, and the template read is in R2, fqdemux / fqtk will output the template read of interests as R2. We deal with this at the nf-core/rnaseq samplesheet level, to ensure that the 'R2' file is treated as a typical single-ended R1.